requirements.txt
diffusers==0.33.1
gdown==5.2.0
huggingface-hub==0.30.2
hydra-core==1.3.2
lmdb==1.6.2
matplotlib==3.10.1
numpy==2.1.2
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.25.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
soupsieve==2.6
sympy==1.13.1
torch==2.6.0
torchdiffeq==0.2.5
torchvision
torchmetrics==1.7.1
tqdm==4.67.1
triton==3.2.0
urllib3==1.26.13
wandb==0.19.9
zipp==3.21.0
